This is the built version of my new website for production.
There are a couple of bugs that I hope to figure out:

1) The "backgroud-clip: text;" property prevents the effect of
a star halfway coloured in black and halfway coloured in gray
in internet explorer (the skills section on narrow browsers).
So I just set the whole star to black in I.E. when it should
be coloured half-black and half-gray

2) The media queries with different image urls for different browser
widths sadly don't work in Firefox and I.E. 11. It works fine in
Edge. I'll try and see if I can fix this, but the end result is not
too bad otherwise.