Fixed a couple of bugs:
1) There was a bug before regarding urls for images in media queries.
It worked in Chrome and Edge, but not in Firefox and I.E. 11. My
solution is to instead create many image tags with the urls for each
image and show them at different screen sizes using media queries,
instead of using the "content: url({insert image source})" property.
I only do this for screen sizes of 992 pixels (px) and below.
There are still such media queries when the screen width is 1100 px
to 992 px but I don't worry about those. It looks ok.
However, this means loading times are slow when the user first loads
the site.

On internet explorer, one of the svgs doesn't load, so the image in the
NBA Records database sub-section of the Projects section will have a
different image in I.E. 11 vs. other browsers.

2) I fixed the issue with the icons at the bottom on mobile phones sized
disproportionately. One of them is an svg, and a separate rule on
svgs caused this.