Tried to change the alignment of icons, and added a disclaimer:
The icons were sized differently, and it showed. I tried changing the
styling to fix this. I added a disclaimer under the skills section
that shows when the chart shows, saying that skills are relative to
my time spent on one skill vs the other. I also added comments in the
CSS files. Next would be to clean up the JS code and add comments in
the JS files.