Changes are as below:
1) There was an issue with the Resume link on the website. It didn't
actually go to the resume on my site.

2) The linkedin icon shown on mobile devices doesn't size properly on
I.E. 11. So on I.E. 11 on mobile devices it wouldn't look good. I
fixed it now

(3 and 4 not applicable here, as this is the built version,
original code here: https://github.com/s4sikdar/react-website)
3) I also tried refactoring the code to create a "CommonUtilities.js"
file, which stores components that are used throughout the project.
It's not much now, but later on I'll continue with it.

4) I started adding some documentation to the JS files. I'll continue
with this.

To do:
1) Continue 4)
2) Check all paragraphs for typos (I know there are a couple)